+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# WIP
+* i : Update librairies
+
 # 1.3.3
 * Fix : tabMover - Auto-close popup with openurl